interceptor creation

+++ Core/ETicaretAPI.Domain/Entities/Product.cs
@@ -11,7 +11,7 @@ namespace ETicaretAPI.Domain.Entities
     {
         public string Name { get; set; } //product ın name propertysi
 
-        public long Price { get; set; } //product ın price propertysi
+        public float Price { get; set; } //product ın price propertysi
         public int Stock { get; set; } //product ın stock propertysi
 
         public ICollection<Order> Orders { get; set; } //bir product birden fazla order da bulunabilir bu yüzden ICollection ile mantıksal yapı kurulur.
